Modernize all Vue webviews; remove JS `unsafe-eval` from webviews (#2449)
This PR eliminates the need for an `unsafe-eval` in src/shared/templates/baseTemplates.ts . This was accomplished by migrating to newer webview building patterns, [which correctly load in Javascript by granting the webview's content security policy](https://code.visualstudio.com/api/extension-guides/webview#content-security-policy), instead of just adding the files directly to the template. 

Most complications have been squashed out, but this does remove the "AWS Lambda" output channel due to issues passing contexts with the new webview model. This results in the following output channel changes:
| Function | Output Channel |
| - | - |
| Invoke remote Lambda | AWS Remote Invocations |
| Delete Lambda | AWS Toolkit |

Most functionality has not been changed (I want this PR to primarily be lift-and-shift rather than a refactor). A key change is in the API Gateway invoker, which replaces a `Map` with a JS object for JSON serialization purposes, caused by not being able to pass context through to the webview's backend.