fix(cloud9): always use patched `got` #2534
Problem:
Default `got` does not work on Cloud9, even after upgrading

Solution:
Always use the 'patched' version. Previously, there was a hard-coded version to
toggle usage of the 'patched' version. The assumption was that the issue was
specific to the VS Code version (not Cloud9), but clearly that's not the case.