Codelens: show without manifest file (#2579)
* Codelens: show without manifest file

Problem: Codelenses (for in code handlers) only appear for files with an associated manifest
file. Lambdas that are downloaded will not have template.yaml file or a
manifest file and therefore no codelens appears, but can still be invoked locally.
Solution: The manifest file was mainly used to find the project root. In
the absence of such file, use the source file as the project root.