fix(ci): disable codecov "patch" check
The codecov "patch" check fails too often, is far too noisy, is not
actionable. Codecov is too unreliable.

At this point, codecov is merely a way to see coverage stats over time,
and to be able to glance at PRs to see if they measurably affected code
coverage (assuming codecov actually works on the PR).