feat(explorer): move CDK to 'Developer Tools' view (#2639)
## Problem

The CDK explorer is basically a copy-paste of the AWS explorer with a few tweaks. This makes it hard to move it around because it has a 'custom' tree provider combined with usage of a coupled base class.

AWSTreeNodeBase is not great for two reasons:

* Directly inherits from vscode.TreeItem
* Developers are encouraged to pack business logic directly in the UI element (!)

The inheritance itself isn't bad, it's just when we also embed logic for user gestures. It makes it really easy to just pipe the node everywhere which we don't want.


## Solution

Separate out the resource ('model') from the tree item in its own interface. I had to add a corresponding data provider as well. Some ideas are still being worked out in terms of ergonomics though the implementation itself is sound.

This PR is the first step towards separating out the model from the tree. CDK was refactored to test some of the ideas.