fix(schemas): yaml schemas sometimes are not applied to cfn/sam files #2671
Problem:
When loading cloudformation files, the validation checks the "Resources" field
in the YAML and see if serverless function exists and then validate those
resources, make sure handlers, codeUri, etc are available before loading in the
schema which doesn't really make sense.

This can cause the cloudformation.load() to fail, resulting in the schema not
getting applied to the cfn/sam yaml files when it should be.

Solution:
Skip one of the validation steps. The net result is that now we only check that
`AWSTemplateFormatVersion` or `Resources` fields exist.