fix(codewhisperer): skip activation when disabled
Problem:
For troubleshooting issues potentially related to codewhisperer, it is
useful for users to be able to fully disable codewhisperer to rule it
out as a possible cause. But codewhisperer activate() is called even
when it is disabled in settings, and this creates various handlers for
numerous vscode events, which could be a potential source of performance
issues or other bugs.

Solution:
Skip codewhisperer activation entirely if it is disabled in
settings.