refactor(cwl): support other kinds of data streams #2719
Problem:
cloudWatchLogs code is written specifically for handing a viewLogStream command
and does not allow for other API calls to implemented in a natural way. 

Currently, the log stream registry in `logStreamRegistry.ts` is built to work
specifically with the `viewLogStream` functionality in mind. One major aspect of
this is how we define `CloudWatchLogStreamData`. This type is defined very
specifically and doesn't allow for a lot of flexibility in terms of working with
other forms of CloudWatchLogs Data that doesn't directly correspond to a stream.

Solution:
Generalize this code and provide easier methods for adding additional
functionality in regards to the CloudWatchLogsAPI. This generalization involves
redefining the `CloudWatchLogsStreamData` type to be a more general
`CloudWatchLogsData` type. The key difference between the two being that the
latter is not specific to handling stream data, but is able to handle the data
from any of the CloudWatchLogs API endpoints.