fix(CodeWhisperer): auto-refresh credentials on Cloud9 #2727
Problem
The Toolkit credentials system does not use a file watcher for detecting credential changes. This specifically affects CodeWhisperer as they never emit an `ExpiredTokenException` which would otherwise trigger a refresh.

## Solution
Force a refresh on errors that are obviously due to expired credentials. We heavily restrict the exact circumstances that this code path can be executed to mitigate risk.

Soon-ish we'll have a file watcher so we can drop a lot of these workarounds.