util: add `guessDefaultRegion` (#2718)
## Problem
Currently, there is not a strong default for what region the user is working with. Internally, it can be difficult to know what region to have defaulted in Wizards or other selections.

## Solution
We can solve this by implementing a heuristic `guessDefaultRegion` function that tracks what regions the user is working with in the explorer and makes a guess on what region the user wants to continue to work with.