feat(credentials): improve "Create Credentials Profile" #2781
Problem:
The template for a new ~/.aws/credentials file is too verbose, and users
tend not to read long explanations. It also clutters up the
out-of-the-box experience so it's harder to see which fields need to
be edited, if starting with a "blank" template.

Solution:
Remove non-essential discursions. Reword for clarity and brevity.