tests: add `testCommand` util (#2831)
## Problem
Not obvious how to test commands without splitting up the logic.
This is only applicable when the command itself requires fakes.

## Solution
Add util function to create testable versions of commands. It's not perfect because VS Code commands are meant to be global and unique whereas testing often requires test versions of things. The implementation is also fragile but acceptable for simple tests.