test: refactor redundant test
This test has a separate describe() to verify a sam
app was created. But this is redundant and the verification
can be done in another test.

This commit removes the old redundant test, but moves
the verification step to an existing test.

Additionally, removal of the old test saves significant
time by reducing a call to createSamApplication() which
indirectly clones a GH repo each time it is called.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>