test: Simplify redundant SAM test
We check the scenario, if a sam app
is created when one already exists with
the same name we expect an error, for
each scenario.
This is not needed and can instead be tested once.

This commit extracts that test in to its own test.

Additionally, the scope of the createSamApplication()
was modified due to it not being accessible by
the new test.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>