fix(test): Flaky codelense registered items
After the previous changes, this test would
fail in CI, but not locally.

It was failing on the line:
assert.ok(projectRoot, 'projectRoot not found')

This was because we would looks for the files found
by the file watcher CodelensRootRegistry right after
calling createSamApplication(). Because of the removal
of redundant createSamApplication() calls in previous
commits, the onDidCreate() event was not executed in
time. This resulted in flakiness.

Force a rebuild of the codelens registry
Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>