skip select dir wizard steps when invoked from template file
Problem: When the uploadLambda command is invoked from a template file,
the wizard prompts for the choice of a zip or a directory and then which
directory. By invoking from the template, we already know the project is
not zipped and the directory should be the parent directory.
Solution: If uploadLambda is invoked from the template file, auto select
the directory option and use the parent directory for the target URI.
Skip both propmts.