feat(dev): add logic to download beta artifacts from a URL (#3038)
## Problem
We have no easy way to share beta builds with users

## Solution
Add codepath to download artifact from a URL. The user is prompted to install the new artifact if it is different from the currently installed extension.

**This codepath is never accessible in a release version**
`src/dev/config.ts` is used to specify the beta URL to ensure that this behavior is dependent on compilation.

### Packaging
A 'config' file has been added to specify development constants. Keeping this as a separate file makes it easier to catch potential leaks and clean the git history of possibly sensitive information. If `betaUrl` is specified then the VSIX will be built with an artificially high version to stop auto-updates e.g. `1.999.0-xxxxxxxx`.