Fix bug if no token in response
When a token is used for the 'final'
page in pagination it will return no
new items and not provide a newer token.
This indicates that the token just used
is the tail.

This commit changes the new token to be
the token we get from the response, but
if there is no token in the response we
use the token from the request.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>