feat(lambda): Copy Url #3036
* chore(#2572): Initial copy lambda URL boilerplate

This commit creates the boilerplate code
which sets up the vscode command to copy
the lambda function url.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

* chore: Logic to get lambda func url

This commit implements the API logic
to retrieve the Lambda Function URL.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

* feat(#2572): Copy Lambda Func URL in context menu

This is the final commit that enables the user
to right click a Lambda function and copy its
URL if it exists.

Additionally, this adds in a new test suite for
the package.json command setup. This ensures all
the aspects of the command registration are configured.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

* Add changelog

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

* simplify object argument properties

This commit changes the argument property
to specify only the information it needs
using 'Pick'.

This helps to simplify the unit tests which
this commit also changes.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

* throw error on cancelled quick pick

When given the option to quick pick
a lambda func url, the user can cancel it.

This commit will now throw an error if the
user cancels the quick pick, instead of
quietly ignoring it.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

* Fix PR comments

This commit:
- Reduces the thoroughness of the extension
  manifest tests. It was a bit too specific.
- Refactors the no lambda function URL message
  and splits the documentation URL in to the
  constants file.
- Fixes the type casting of a Lambda mock to
  be more accurate, though not perfect due
  to limitations.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

* Add quickpick test

This refactors the quick pick functionality
so it can be tested. And also adds in the test.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

* chore: Remove unnecessary test

The lambda client test is not
necessary currently due to not
having a defined way to test SDKs.

This commit removes the test that
tests the SDK and slightly refactors
some code with the removal of this.

Also this removes the parameter for
passing in a SDK since it is no longer
needed.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>