fix(stepFunctions): #2922 - fix offline sfn visualizer
This fix allows for a graceful fallback when internet connectivity isn't available to a c9 or vscode host but where the required
files are available in the local filesystem cache (either by previous automatic caching, or by manual user intervention).
Implementation avoids confusing the return type of 'updateCache' or 'updateCachedFile' and instead only checks for local files
after a updateCache call fails.