chore: Reload window when disabling lsp
Calling stop() on the LSP client through
vscode resulted in a failed LSP JSON RPC
'shutdown' call. At this point I don't
know why that fails. When debugging the
Rust `shutdown` api is never breakpointed.

As a result of this, this will reload the
window if the user selects 'Yes' in the prompt.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>