fix: "Upload Lambda" from template.yaml asks for directory #3043
Problem: When "Upload Lambda" is invoked from a template.yaml file in the vscode
file explorer, the wizard prompts for the choice of a zip or a directory and
then which directory. These steps are redundant.

Solution: When "Upload Lambda" is invoked from a template.yaml,
skip the wizard steps that ask to choose a directory.