fix: unit test case failing on windows machine for codeWhisperer truncation lines
Problem: Unit test cases of truncation lines for codeWhisperer were failing on windows OS.

Solution: We have to use vscode.uri.file() rather than vscode.Uri.parse().
The difference between Uri#parse and Uri#file is that the latter treats the argument
as path, not as stringified-uri. E.g. Uri.file(path) is not the same as Uri.parse(‘file://’ + path) because the path might contain characters that are interpreted (# and ?).