fix(codecatalyst): cache results of `verifySession` (#3113)
## Problem
verifySession is called every time a new client is created

## Solution
Cache verifySession based off the access token

### Testing
The way the CodeCatalyst client is written makes it difficult to test in a sustainable way. Something similar to #2917 needs to be done to reduce the amount of custom logic in each service client.