WIP:upload lambda sort regions from appJson to top
Problem: We have metadata information about downloaded lambdas stored in
an application.json file, but in the upload lambda flow we do not see a
connection for region.
Solution: Add a description to the regions that are found in the
application.json file and move those regions to the top of the list.