refactor: Use cwl filter api call instead of get events
Problem

Prior to this commit we had a separate function which
did the same thing as filterLogEventsFromUriComponents
except it used a different API call. This separate
function was not necessary.

Solution

Remove the old function, replace uses of it
with filterLogEventsFromUriComponents. Also
slightly modify the filterLogEventsFromUriComponents
to filter the log stream name if it is given.

This reduces the amount of code.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>