chore: Add cancellable vscode window manager
Problem

We can create a vscode popup window that is cancellable,
or closes itself after a certain timeout. The issue is
when we make multiple requests that create a cancellable
window they tend to stack upon each other. This creates
too many instances on the screen. Ultimately we only care
about the latest window.

It is not easy to manage these windows since one place
may create the message, but not know when the process
is done so that it can be fully cancelled.

Solution

This commit creates a CancelablleMessageManager that
holds a global state which can be used to create, updated,
and cancel cancellable messages in a way that makes it
easier to avoid duplication issues.

Additional

This commit makes use of the new manager by first creating
the message in one area of the code, then having a separate
function update the message as it runs, finally the message
is cancelled when we know everything is done.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>