auth: add login/cancel explorer buttons
Problem: The AWS explorer should show a more obvious status before and
during logging in. The actions available should also be more explicit
and easy to find.

Solution: When the user is not logged in, show a 'Login' button. If
authentication is in process, show a 'Cancel' button. Show a button to
switch profile.