feat(stepfunctions): update amazon-states-language-service #3160
Problem:
Customers are not able to validate their state machines definition with new
intrinsic functions and new Map state attributes. This behaviour is not
consistent with the launch of the new distributed Map state, and new intrinsic
functions

Solution:

https://github.com/aws/amazon-states-language-service/pull/90
https://github.com/aws/amazon-states-language-service/pull/86
added validation for both distributed Map fields and new intrinsic functions.
Bump the package version to `^1.8.0` to pull in this change.

Testing:
Confirmed that Map field values and new intrinsic functions were correctly
validated when creating a new state machine, when bumping the version to
`^1.8.0`.