chore: Add function to get content with eTag
The http resource fetcher class did not provide
an option to give an eTag and attempt to fetch
fresh content.

This commit adds in a function that users can call
with an eTag, it will then return the response which
would be new content with a new eTag, OR if the given
eTag is the same as the remote it will give back
undefined content.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>