chore: Have cfn use the new sam schema
- Problem

For our doctype suport we use different schemas
for cfn and sam. But now there is an all-in-one
sam schema that works for cfn as well.

- Solution

Have cfn and sam use the new sam schema. This
requires us to update our url content fetching
+ caching mechanisms to use e-tags to determine
stale content instead of using versions as we
were previously doing.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>