fix(auth): add extra notification to SSO login flow (#3148)
## Problem
Logging in via the browser is currently too easy. Section 5.4 of RFC 8628 suggests that implementations should make it clear to users which device is requesting access at all times.

## Solution
Add an additional notification containing the user verification code before opening the browser. Users must now enter in this code in the browser whereas before it was done automatically. The user code is automatically copied to the clipboard to make things slightly easier.

## Relevant changes
* Add notification to SSO portal login flow
* Globally fake `vscode.window` for unit tests