chore: Create Log Stream Code Lens Provider
This commit creates a new code lens provider
class that will create code lenses on a CWL
document of Log Stream events.

A CWL Document can consist of contiguous log
streams. This will create a code lens at the
start of each log stream which when clicked
will open up a new document that has all the
events of that log stream.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>