tests: remove `Window.vscode()` and related logic (#3182)
## Problem
Writing wrappers for the sole purpose of making an implementation easier to test is not a sustainable approach.

## Solution
* Remove `Window.vscode()` and `FakeWindow`
* Extend `TestWindow` to cover the same scenarios that `FakeWindow` was used for

### Important Notes
This PR uses Proxy heavily in order to make "testable" versions of native vscode objects. Some things like `showInformationMessage` and `withProgress` do not have an API that can be manipulated programmatically. For these methods, we have to create in-mem representations that can be tested. These currently live in `src/test/shared/vscode/message.ts`.