file explorer: add upoloadToAws command
Problem: The toolkit can upload files and folders to Lambda and S3, but
the commands are only available from the AWS Explorer or the Command
Palette.

Solution: Add a context menu item to all files and folders that prompts
for the resource to upload to. Initiate the apprpriate wizard passing
the file/folder. Current options are Lambda and S3.