chore: Option to cancel message on filter logs
When we call the filter logs function we want to be
able to choose if the vscode cancel message close itself
when the underlying fetch is completed.

Additionally, a function that explicitly calls the filter
log events function previously kept the cancel message window
open. But now it explicitly tells the filter log events
method to close the window.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>