tests(codecatalyst): start basic integration suite (#3207)
## Problem
We want to ensure that Code Catalyst apis work as expected. If only we had integration tests that could do this.

## Solution
This commit creates integ tests that test the CC APIs
this extension uses:
* creating a cc project
* creating a cc dev env
* connecting to the dev env through ssh

One thing to note is that the builderid/sso configuration is required to be setup prior to this test running (this test will fail until that component is implemented in our CI, but is being worked on separately).