feat(syntax): use SAM-provided schema instead of goformation #3163
Problem:
The http resource fetcher class did not provide an option to give an
eTag and attempt to fetch fresh content.

Solution:
Add a function that users can call with an eTag, it will then return the
response which would be new content with a new eTag, OR if the given
eTag is the same as the remote it will give back undefined content.

Problem:
For our doctype suport we use different schemas for cfn and sam. But now
there is an all-in-one sam schema that works for cfn as well.

Solution:
Have cfn and sam use the new sam schema. This requires us to update our
url content fetching + caching mechanisms to use e-tags to determine
stale content instead of using versions as we were previously doing.


* test: integ test for getDefaultSchemas() happy path

    The getDefaultSchemas() function downloads a schema
    if it doesn't exist yet in cache.

    This test uses the telemetry data to determine if
    the cache is used when a request for a schema is made,
    otherwise downloading it.

    - Additionally, this updates the schema assertion function
      to allow the user to choose the event occurence they want
      to assert against.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>