feat(IDE-10199): prompt existing builder id signout on new builder id
Problem:

When a user is already signed in to their Builder ID, but
they want to use a different one, they are not able to
use a different ID.

This is confusing since nothing happens when they try to
add another one.

Solution:

When attempting to add a new Builder ID but one already
exists, this will prompt the user to choose if they wish
to sign out of the existing one, and will then complete
the builder id verification process.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>