feat(IDE-9298): Dev Env Project Quick Pick selector
Problem:

When using the Create Dev Env webview we used
a drop down menu to provide all possible Projects.

The problem is that querying all Projects takes time
and nothing would show on the screen for a while.

Additionally, if the user had many projects from
multiple spaces it was not easy to find

Solution:

Instead of a dropdown menu use an async quick pick
which will continuously load new Projects as they
come in the the paginated query for all Projects.

- Along with using a quick pick, the user can easily
filter the Projects using the quick picks built in
filtering mechanism.

- Upon initial load of the webview, the fetching of
  projects is started immediately. This will improve UX.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>