fix(codecatalyst): Disable new branch name 3p repo
Problem:

When creating a new Dev Env in the webview, users
have the option to create a new branch name from a
linked repo.

The issue though is we are not able to create a new
branch name for third party (github) repos.

Solution:

Do not allow input of a new branch name when the selected
branch is from a third party repo.

Fixes IDE-10216

- This allows input from the user only when necessary
- Otherwise input is disabled
- Verifies if new branch is allowed once a user selects
  an existing branch

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>