fix: disable existing profile check when adding profile
Problem:

When adding an auth profile, there is a possible race
condition where the global store (memento) is not
synchronized on a following read.

The scenario we were hitting was signing out of
a builder id, then signing in with a new one. It
would fail since the old builder id would still be
in the memento.

Soltion:

Do not check if the profile already exists,
just overwrite it.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>