fix(s3): set file system provider to be case sensitive
Problem:

When the s3 file system provider was registered it
was not set to be case sensitive. So paths like
/aws/Test/temp.json and /aws/test/temp.json were
seen as the same thing.

Solution:

Set the case sensitive flag when registering the
file system provider so that it will uniquely
identify files with the same text but different
case sensitivity.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>