feat(cwl): Remember previous selection for search/filter
Problem:

To search a CWL log group you must enter a date
range, and an optional string to filter the
event text.

Once openend there are buttons in the editor/title
that are the calendar and magnifying glass.

If you click on one of these you can re-adjust their
value through a quickpick/input prompter.

The problem is that if users want to readjust their
date range using the calendar button at the top
right of the document, it does not remember their
previous selection.

Solution:

When the user attempts to re-adjust their previous selection,
autofill the previous values in the prompter so they
can see what they previously chose.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>