fix(codecatalyst): invalidate connections refreshed by MDE (#890)
## Problem
Users who decide to enable CW in a dev environment might have their credentials refreshed by MDE if they reconnect some time later. These new credentials do not have CW scopes, causing confusing auth issues.

## Solution
We write the scopes to disk with the bearer token credentials so all we have to do is check if the scopes align with our current state. If not, invalidate the token.

This behavior likely will be added to all connections and not just CC + builder ID in a later PR.