fix(codecatalyst): dev env error messages not being displayed
Problem:

The following issue was started due to a change needed
for Velox, but this will be able to apply to all AWS
services.

When certain exceptions are thrown by AWS services
we want them to be displayed directly to the user in
a vscode window.

Currently we only display Toolkit errors to the user,
otherwise only logging the error to the logs.

There are some AWS errors that we want to show to the
user, but on top of that AWS errors can be wrapped
inside a Toolkit error.

Solution:

When deciding which error message to show to users:

- Show the exact error message if it is an AWS error
  and is in our defined list of 'prioritizied' errors.
- If we get a Toolkit error, check if it wraps a prioritized
  error, extract it, then display that.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>