fix(codecatalyst): No solution for 'connection is invalid or expired'
Problem:

When using an MDE + CodeCatalyst, if the BuilderId connection
became expired, on attempts to use the CodeCatalyst client
users would get an error 'Connection is invalid...' with no
obvious way to resolve it.

Solution:

Provide the user the ability to re-authenticate their BuilderId.

This commit verifies the BuilderId connection is valid before
setting up the CodeCatalyst client. If it is not valid, users
will be prompted to login again. They can click the login button
on a pop up window, and be redirected through the BuilderId
authentication process again.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>