fix(codecatalyst): Notify ssh extension min version required
Problem:

When a user attempted to start a dev env, it failed
since the ssh extension version was too old.

Solution:

Enforce a minumum version of the ssh extension.
If the version is unsupported we will notify the user
through an error message.

This solution can be used by any future extension
as well.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>