fix(codewhisperer): preserve CC+CW token on dev env reconnect (#891)
## Problem
Reconnecting to a dev environment after adding CW scopes always results in the connection starting out as expired. This is because MDE writes over our token making it unusable.

## Solution
Write the token to a different location when using CW scopes within a dev environment. This is done by changing the token provider ID. MDE uses codecatalyst but we'll use our own identifier.