feat(cwl): jump to search log group from log stream quick pick
Problem:

By default when a user clicks on a Log Group in the explorer
they will be shown a quick pick of only Log Streams.

If they want to search the whole Log Group, they have to
exit out of it, right click, then select search log group.

Solution:

Add a quickpick item as the first item of the Log Stream
selection. This item if chosen will redirect them to the
Search Log Stream quick pick flow.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>