fix(endpoints): guard against remote call hanging #3341
Problem:
On startup, Toolkit tries to fetch https://idetoolkits.amazonwebservices.com/endpoints.json
in EndpointsProvider#load. If that URL is is blocked by firewall, the toolkit is
blocked and unusuable. #3338

Solution:
First load local endpoints.json then dynamically update with the result of
remote call once it receives a response.