fix(codecatalyst): do not force-close workspace in Cloud9 devenv
Problem:
Since eb380556b6b4, Toolkit actively exits if the CC devenv is not
a "vscode" devenv. But this is incorrect when running in the context of
a Cloud9 devenv.

NOTE: executeCommand('workbench.action.remote.close') has no effect in
Cloud9, but this the code should be corrected anyway.

Solution:
Skip the condition in Cloud9.