fix(codecatalyst): e2e test updateDevEnv failing
For some reason the cc backend did not receive at
dev env id from our e2e test updateDevEnv api call.

This commit verifies we are giving a dev env id
before sending the api request.

Additional:

We used the variable 'global' before it could
be properly initialized and due to the order of
execution this variable was accessed before it
could be properly initialized.

To fix this it was moved inside a 'before' block
which will ensure it gets executed AFTER all the
pre-test hooks are fun, which includes initializing
the global variable

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>