Fix CW connection not refreshing on vscode startup
This PR also fixes some other issues:
1. Clear the inlineCompletionState on focus change so that
they should still be able to invoke CW when they invoked CW and then
switched to another app and switched back.
2. put activateCodeWhipserer() to be before activateAwsExplorer()
since the refreshStatusBar() and showReferenceLog() commands are used
within activateAwsExplorer() when the connection is changed. These
 commands are not registered until activateCodeWhisperer() has been called.
3. Remove unnecessary call to refreshCodeWhisperer(), this only needs to be
called when there's a connection change.