fix(auth): invalid sso tokens being loaded
Problem:

At a certain point in time sso tokens were became invalid.
We stored those tokens in cache and eventually they will
throw errors to the user when the are attempted to be
refreshed.

Solution:

When the cache is being setup delete any invalid files.
Invalid files are files that were created before
a certain date.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>