function to add credentials with all data at once
Problem:

Our code was initially designed with adding credentials
through the vscode api flow in mind. This does not work
with the new method of using a webview

Solution:

Create a new function that allows the credentials data
to be submitted at once where it will validate the data,
add it to the credentials file, then possibly create a
connection.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>