method to authenticate raw credentials data
Problem:

When I have raw credentials data and want to
authenticate it with our existing authentication
methods it is not easy to set up since the data
must exist in the store.

The way to add it to the store is to add it to
the credentials file, or use an env var so that
it will be picked up by an auth credentials provider.

Solution:

Create a simple auth credentials provider that is
easy to set up with only the necessary data.

Notes:

- This creates methods in the Auth class to easily
  add and remove the credentials provider

- This commit adds a method throwIfCantAuthenticateCredentials()
  which uses this functionality to easily test if credentials
  can be authenticated while hiding auth provider setup

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>