throttle connection change event firing
We listen to multiple connection change events
and then trigger a full rerender of all states.

Due to listening to multiple events that tend
to trigger at the same time, we change events
and then trigger a full rerender of all states
multiple times.

SOLUTION:

throttle the event firing so that it only happens
with a certain delay.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>