fix(codewhisperer): extra space at the beginning of ghost text #3486
Problem:
When user types with SPACE at certain speed, both the service and user will add extra SPACE after the CodeWhisperer invocation position, resulting in multiple extra space at the start of the suggestion ghost text, which breaks the indentation of the current file.  

The root cause is an old mechanism that moves the invocation start position if the user typeahead is while-space characters.

Solution:
Remove the old start position moving mechanism. Make sure the space is included in the typeahead computation.