delayed rendering for auth loading
Problem

In the UI a content window (cc, cw, ...) can have multiple
auth forms within it.

These auth forms need to communicate with the backend
so they can update to the latest state (check if the
auth is already connected)

Since we want the auth form to be independent from its parent
component we need to rely on the auth form to resolve
its own status when it is created.

The real problem is that one version of the auth form is
displayed to the user, but the UI flickers as it updates
the form to the latest state.

Solution

Have the auth form make all its backend calls, but have
the form emit an event to indicate once it is done loading.

We can then delay displaying the auth form in its parent
component and only show once it is fully ready.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>